+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [1.3.4] - 2017-09-xx
 
 - Added variable $menu-offset-nav-bg-color to the OffsetNav.
+- Added variable $overlay-bg-color to the Overlay. 
 
 
 ## [1.3.3] - 2017-09-14